# docs/api/openapi/openapi.yaml
...
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭认知健康日记 API",
    "version": "1.0"
  },
  "paths": {
    "/v1/auth/email-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Email Code",
        "operationId": "send_email_code_v1_auth_email_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/password-reset": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Password Reset",
        "operationId": "password_reset_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_v1_accounts_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_v1_accounts_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subjects": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Subjects",
        "operationId": "list_subjects_v1_subjects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Subject",
        "operationId": "create_subject_v1_subjects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subjects/{subjectId}": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Patch Subject",
        "operationId": "patch_subject_v1_subjects__subjectId__patch",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subjectid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subjects/{subjectId}/link": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Unlink",
        "operationId": "unlink_v1_subjects__subjectId__link_delete",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subjectid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subjects/{subjectId}/links": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Links",
        "operationId": "list_links_v1_subjects__subjectId__links_get",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subjectid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/session/subject": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Lock Subject",
        "operationId": "lock_subject_v1_session_subject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSubjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteAccountIn": {
        "properties": {
          "confirmText": {
            "type": "string",
            "title": "Confirmtext"
          }
        },
        "type": "object",
        "required": [
          "confirmText"
        ],
        "title": "DeleteAccountIn"
      },
      "EmailCodeIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "scene": {
            "type": "string",
            "pattern": "^(REGISTER|PASSWORD_RESET)$",
            "title": "Scene"
          }
        },
        "type": "object",
        "required": [
          "email",
          "scene"
        ],
        "title": "EmailCodeIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginIn": {
        "properties": {
          "account": {
            "type": "string",
            "title": "Account"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "client": {
            "type": "string",
            "pattern": "^(ELDER_APP|FAMILY_APP)$",
            "title": "Client"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "account",
          "password",
          "client"
        ],
        "title": "LoginIn"
      },
      "LogoutIn": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          }
        },
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "LogoutIn"
      },
...
      "PasswordResetIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "email",
          "code",
          "newPassword"
        ],
        "title": "PasswordResetIn"
      },
...
      "RefreshIn": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          }
        },
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "RefreshIn"
      },
      "RegisterIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Code"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "gender": {
            "type": "string",
            "pattern": "^(MALE|FEMALE)$",
            "title": "Gender"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "code",
          "password",
          "gender"
        ],
        "title": "RegisterIn"
      },
...
      "SessionSubjectIn": {
        "properties": {
          "subjectId": {
            "type": "string",
            "title": "Subjectid"
          }
        },
        "type": "object",
        "required": [
          "subjectId"
        ],
        "title": "SessionSubjectIn"
      },
      "SubjectCreateIn": {
        "properties": {
          "elderName": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Eldername"
          },
          "bindCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bindcode"
          },
          "ageBand": {
            "type": "string",
            "title": "Ageband"
          },
          "dialect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect"
          },
          "hearingIssue": {
            "type": "boolean",
            "title": "Hearingissue",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "elderName",
          "ageBand"
        ],
        "title": "SubjectCreateIn"
      },
      "SubjectPatchIn": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "hearingIssue": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hearingissue"
          },
          "dialect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect"
          }
        },
        "type": "object",
        "title": "SubjectPatchIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}